Rename schema -> pubspec_schema
