+++ packages/flutter_tools/test/android/gradle_test.dart
@@ -18,7 +18,7 @@ import 'package:process/process.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/schema.dart';
+import '../src/pubspec_schema.dart';
 
 void main() {
   Cache.flutterRoot = getFlutterRoot();
