+++ packages/flutter_tools/test/asset_bundle_package_fonts_test.dart
@@ -15,7 +15,7 @@ import 'package:flutter_tools/src/cache.dart';
 
 import 'src/common.dart';
 import 'src/context.dart';
-import 'src/schema.dart';
+import 'src/pubspec_schema.dart';
 
 void main() {
   String fixPath(String path) {
