+++ packages/flutter_tools/test/flutter_manifest_test.dart
@@ -12,7 +12,7 @@ import 'package:flutter_tools/src/flutter_manifest.dart';
 
 import 'src/common.dart';
 import 'src/context.dart';
-import 'src/schema.dart';
+import 'src/pubspec_schema.dart';
 
 void main() {
   setUpAll(() {
