+++ packages/flutter_tools/test/ios/xcodeproj_test.dart
@@ -17,7 +17,7 @@ import 'package:process/process.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/schema.dart';
+import '../src/pubspec_schema.dart';
 
 const String xcodebuild = '/usr/bin/xcodebuild';
